Extend LanServers to all addresses
Adds ways to send multicast packages (a) for IPv6 addresses and
(b) for all local interfaces.

As it happens, "just" sending a multicast packge will, by default
make it's "source" (almost always) the primary IP of the machine.
But if the server is listening on multiple (or even all) interfaces,
not eveyone will get the memo. So we iterate all local interfaces
(that can do multicast and are actually up, but ignore localhost,
because what's the point?) and send messages _from_ these addresses.
This is a bit involved due to how interfaces/addresses/sockets
work in Java.